refactor(voyage): le gel des jambes et les commodités du parcours sortent d'app.js
`figerJambe`, `pinLegsForVolume` et `journeyCarriedCommodities` ne rendent
rien : elles lisent le parcours, décident quoi geler, et persistent. Elles
rejoignent `voyage-donnees.ts`, où leurs commentaires les attendaient déjà —
trois blocs orphelins laissés par la PR #144, sans fonction en dessous.

C'est le préalable des deux migrations qui suivent : les QUATRE portes vers
le gel (Trajets, Manifeste, Corrections, Commodités) appellent
`pinLegsForVolume` avant d'écrire un volume, et deux de ces quatre vont
emménager dans l'arbre. `journeyCarriedCommodities` est lue par le récap de
voyage ET par le board des Commodités.

`figerJambe` gagne une garde `if (!etat.JOURNEY) return false` que la version
d'app.js n'avait pas : elle déréférençait `etat.JOURNEY.legs[i]` sans la
poser. Ses deux appelants la garantissaient, le type ne le disait pas.

`legsToPin` quitte les imports d'app.js — plus aucun appelant.
